Superclasses and Interfaces
The program is now capable of displaying the classes superclasses as
well as the interfaces.  The call its self to get the superclass what
quite easy but the formatting of the output took a bit of time.  I think
i will use the same formatting in the methods to come as it seems to
give a nice readable output.