Methods and Constructors
Got the inspect Methods and inspect Constructors working.  The
constructors method turned out to be quite short due to the fact that
you only looking for a singular thing.  The inspect Methods method on
the other hand had to not only get the name of the methods but also
their type and modifiers and parameters.  tried to use appropriate white
space for the output so that the data is readable.