Final submit for core changes / small refactor
Added the final functional aspects of this program.  Performed a minor
refactoring to make some of the variable names more informative.  Tested
the whole program with the test driver and the results seemed to be what
was required in the specs.  I was able to reuse one of the original
functions almost completely only having to add more output so that it
matches the rest of the program i wrote.